Move IArgConsumer from *.helpers.arguments to *.argument
