+++ src/api/java/baritone/api/command/Command.java
@@ -21,7 +21,7 @@ import baritone.api.IBaritone;
 import baritone.api.utils.Helper;
 import baritone.api.utils.IPlayerContext;
 import baritone.api.command.exception.CommandException;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 
 import java.util.Collections;
 import java.util.List;
