+++ src/api/java/baritone/api/command/argument/ICommandArgument.java
@@ -17,7 +17,6 @@
 
 package baritone.api.command.argument;
 
-import baritone.api.command.helpers.arguments.IArgConsumer;
 import baritone.api.command.argparser.IArgParser;
 import baritone.api.command.exception.CommandInvalidTypeException;
 import net.minecraft.util.EnumFacing;
