+++ src/api/java/baritone/api/command/datatypes/IDatatype.java
@@ -18,7 +18,7 @@
 package baritone.api.command.datatypes;
 
 import baritone.api.command.argparser.IArgParser;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 import baritone.api.command.exception.CommandException;
 
 import java.util.stream.Stream;
