+++ src/api/java/baritone/api/command/helpers/tabcomplete/TabCompleteHelper.java
@@ -21,7 +21,7 @@ import baritone.api.BaritoneAPI;
 import baritone.api.Settings;
 import baritone.api.event.events.TabCompleteEvent;
 import baritone.api.utils.SettingsUtil;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 import baritone.api.command.manager.ICommandManager;
 import net.minecraft.util.ResourceLocation;
