+++ src/main/java/baritone/command/BaritoneChatControl.java
@@ -29,7 +29,7 @@ import baritone.api.utils.SettingsUtil;
 import baritone.api.command.argument.ICommandArgument;
 import baritone.api.command.exception.CommandNotEnoughArgumentsException;
 import baritone.api.command.exception.CommandNotFoundException;
-import baritone.command.helpers.arguments.ArgConsumer;
+import baritone.command.argument.ArgConsumer;
 import baritone.api.command.helpers.tabcomplete.TabCompleteHelper;
 import baritone.api.command.manager.ICommandManager;
 import baritone.command.argument.CommandArguments;
