+++ src/main/java/baritone/command/defaults/BuildCommand.java
@@ -24,7 +24,7 @@ import baritone.api.command.datatypes.RelativeBlockPos;
 import baritone.api.command.datatypes.RelativeFile;
 import baritone.api.command.exception.CommandException;
 import baritone.api.command.exception.CommandInvalidStateException;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 import net.minecraft.client.Minecraft;
 
 import java.io.File;
