+++ src/main/java/baritone/command/defaults/ComeCommand.java
@@ -22,7 +22,7 @@ import baritone.api.pathing.goals.GoalBlock;
 import baritone.api.command.Command;
 import baritone.api.command.exception.CommandException;
 import baritone.api.command.exception.CommandInvalidStateException;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 import net.minecraft.entity.Entity;
 import net.minecraft.util.math.BlockPos;
