+++ src/main/java/baritone/command/defaults/CommandAlias.java
@@ -19,7 +19,7 @@ package baritone.command.defaults;
 
 import baritone.api.IBaritone;
 import baritone.api.command.Command;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 
 import java.util.Collections;
 import java.util.List;
