+++ src/main/java/baritone/command/defaults/ExploreCommand.java
@@ -22,7 +22,7 @@ import baritone.api.pathing.goals.GoalXZ;
 import baritone.api.command.Command;
 import baritone.api.command.datatypes.RelativeGoalXZ;
 import baritone.api.command.exception.CommandException;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 
 import java.util.Arrays;
 import java.util.List;
