+++ src/main/java/baritone/command/defaults/FindCommand.java
@@ -22,7 +22,7 @@ import baritone.api.utils.BetterBlockPos;
 import baritone.api.command.Command;
 import baritone.api.command.datatypes.BlockById;
 import baritone.api.command.exception.CommandException;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 import net.minecraft.block.Block;
 
 import java.util.ArrayList;
