+++ src/main/java/baritone/command/defaults/GoalCommand.java
@@ -25,7 +25,7 @@ import baritone.api.command.Command;
 import baritone.api.command.datatypes.RelativeCoordinate;
 import baritone.api.command.datatypes.RelativeGoal;
 import baritone.api.command.exception.CommandException;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 import baritone.api.command.helpers.tabcomplete.TabCompleteHelper;
 
 import java.util.Arrays;
