+++ src/main/java/baritone/command/defaults/InvertCommand.java
@@ -24,7 +24,7 @@ import baritone.api.process.ICustomGoalProcess;
 import baritone.api.command.Command;
 import baritone.api.command.exception.CommandException;
 import baritone.api.command.exception.CommandInvalidStateException;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 
 import java.util.Arrays;
 import java.util.List;
