+++ src/main/java/baritone/command/defaults/SelCommand.java
@@ -35,7 +35,7 @@ import baritone.api.command.datatypes.RelativeBlockPos;
 import baritone.api.command.exception.CommandException;
 import baritone.api.command.exception.CommandInvalidStateException;
 import baritone.api.command.exception.CommandInvalidTypeException;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 import baritone.api.command.helpers.tabcomplete.TabCompleteHelper;
 import baritone.utils.IRenderer;
 import net.minecraft.init.Blocks;
