+++ src/main/java/baritone/command/defaults/ThisWayCommand.java
@@ -21,7 +21,7 @@ import baritone.api.IBaritone;
 import baritone.api.pathing.goals.GoalXZ;
 import baritone.api.command.Command;
 import baritone.api.command.exception.CommandException;
-import baritone.api.command.helpers.arguments.IArgConsumer;
+import baritone.api.command.argument.IArgConsumer;
 
 import java.util.Arrays;
 import java.util.List;
